backend/requirements.txt
asgiref==3.7.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
Django==5.0.3
django-jalali-date==1.1.0
django-render-partial==0.4
fonttools==4.47.0
graphviz==0.20.1
gunicorn==21.2.0
jdatetime==4.1.1
joblib==1.3.2
kiwisolver==1.4.5
neat-python==0.92
packaging==23.2
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
PyTMX==3.32
pytz==2024.1
regex==2023.12.25
six==1.16.0
sorl-thumbnail==12.10.0
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
whitenoise==6.6.0
environs==9.5.0
psycopg2==2.9.9
dj-database-url==2.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1 
django-allauth==0.61.1
dj_rest_auth==6.0.0
Unidecode==1.3.8
unicode-slugify==0.1.5 